Relay: extract operation persisters
I am choosing a different strategy here: we are not using the persister yet so I want to extract it and add FS implementation. This way we'll have it ready for the future when we decide to actually use it. I used the FS persister in Relay example for demo purposes.

kiwicom-source-id: 7332569e01bef2a2aa948327f4abbdc2e3465c38