Flowconfig: enable 'untyped-type-import' lint
The most common problem was in generated files. I simply disabled the lint for Relay generated files at this moment so we can enable the lint and people can benefit from it. Later on, we should fix the generated files. This is however not so straightforward in all the cases.

kiwicom-source-id: f7b9275e6f58c73d77e7ee3f2cf9496efca9050c